+++ CHANGELOG.md
@@ -23,6 +23,12 @@ _This sections lists changes committed since most recent release_
 **Fixed:**
 - Bug where a video unit's thumbnail was not being assigned to `unit.thumbnail` on the results page; This was preventing a video unit's thumbnails from being listed in its download popup.
 
+
+# [5.1.1](https://github.com/IIP-Design/content-commons-client/compare/v5.1.0...5.1.1)(2020-07-29)
+
+**Fixed:**
+- Issue causing some video downloads to fail. Use filename and not title for download filename  
+
 # [5.1.0](https://github.com/IIP-Design/content-commons-client/compare/v5.0.0...5.1.0)(2020-07-15)
 
 **Added:**